refactor detail view
